+++ friends_server/src/main/kotlin/com/friends/security/service/AuthService.kt
@@ -21,9 +21,7 @@ import com.friends.security.CheckEmailResponseDto
 import com.friends.security.CheckNicknameResponseDto
 import com.friends.security.OAuth2LoginDto
 import com.friends.security.RegisterRequestDto
-import com.friends.security.securityException.DuplicateNewPasswordException
 import com.friends.security.securityException.EmailDuplicateException
-import com.friends.security.securityException.EmailNotFoundException
 import com.friends.security.securityException.InvalidNicknameException
 import com.friends.security.securityException.InvalidPasswordException
 import com.friends.security.securityException.InvalidRefreshTokenException
@@ -145,7 +143,7 @@ class AuthService(
         val nickname = registerRequestDto.nickname
         val password = registerRequestDto.password
         // 패스워드 유효성 검사
-        if (!validatePassword(password!!)) {
+        if (!Member.validatePassword(password!!)) {
             throw InvalidPasswordException()
         }
         return Member.createUser(
@@ -153,20 +151,6 @@ class AuthService(
             email = email,
             password = passwordEncoder.encode(password),
         )
-    }
-
-    fun validatePassword(password: String): Boolean {
-        val lengthRegex = Regex(".{8,20}") // 길이 제한
-        val lowerCaseRegex = Regex(".*[a-z].*") // 소문자 포함
-        val digitRegex = Regex(".*[0-9].*") // 숫자 포함
-        val specialCharRegex = Regex(".*[!@#\$%^&*(),.?\":{}|<>].*") // 특수문자 포함
-        val noWhiteSpaceRegex = Regex("^[^\\s]*\$") // 공백 금지
-
-        return lengthRegex.matches(password) &&
-            lowerCaseRegex.containsMatchIn(password) &&
-            digitRegex.containsMatchIn(password) &&
-            specialCharRegex.containsMatchIn(password) &&
-            noWhiteSpaceRegex.matches(password)
     }
 
     fun validateNickname(nickname: String): CheckNicknameResponseDto {
@@ -226,34 +210,5 @@ class AuthService(
             atRtService.getAccessToken(refreshToken)?.let { atRtService.deleteAccessToken(it) }
             atRtService.deleteRefreshToken(refreshToken)
         }
-    }
-
-    @Transactional
-    fun resetPassword(
-        emailAuthToken: String,
-        newPassword: String,
-    ) {
-        // emailAuthToken 검증
-        if (!jwtService.validate(emailAuthToken)) {
-            throw InvalidTokenException()
-        }
-        val emailFromToken = jwtService.getClaim(emailAuthToken, "email", String::class.java) ?: throw InvalidTokenException()
-
-        // 유저 존재 여부 확인
-        val member = memberRepository.findByEmail(emailFromToken) ?: throw EmailNotFoundException()
-
-        // 패스워드 유효성 검사
-        // 적절한 비밀번호 패턴인지 검사
-        if (!validatePassword(newPassword)) {
-            throw InvalidPasswordException()
-        }
-        // 새 비밀번호와 기존 비밀번호가 같은지 검사
-        if (passwordEncoder.matches(newPassword, member.getPassword())) {
-            throw DuplicateNewPasswordException()
-        }
-
-        // jwt 에서 email 을 추출하고 해당 email 을 가진 사용자의 비밀번호를 변경합니다.
-        member.updatePassword(passwordEncoder.encode(newPassword))
-        memberRepository.save(member)
     }
 }
